tone down all the debug output

+++ config.hpp
@@ -5,7 +5,7 @@
 
 
 #define PointScalar double
-typedef float  RealScalar;
+typedef double  RealScalar;
 //#define  RECURSIVE_MULT
 //#define CHECK_CONNECTIVITY
 //#define BE_FAST
